Add a small query language for states.
A new message (UState::Query) contains a string query like

(state = "critical" or host = "winer") or service =~ "% daemon"

Returns a protobuf message containing a repeated list of States
matching the query. ~700 queries/sec on my core 2 duo. The Message
protobuf is now repeated: states.

Also includes improved error handling; exceptions are relayed in
Message#error when Message#ok == false, and re-raised on the Client.

The query language is a treetop parser generator. Yes, this is a program
which converts a program specifying a program into a program parsing a
language, which generates a program, which generates SQL. Yo dawg.