Add a basic functioning dashboard.
The dashboard is a small sinatra app that uses UState::Client to request
states from a server and display them in a nice dashboard. Features
configuration through an instance-evaled config.rb, sass rendering, a
small library of helper tags, overridable views, and a pluggable
controller system. I have not tested the entire library of helpers, and
caching is wanting. Fast enough for now, though.