States now have a :once boolean.
:once states are transient. They fire Index#on_state and #on_state_once,
but do *not* update the index. They can be used for events which are
instantaneous; instead of sending {state: error} and {state: ok}, send
{state: error, once:true}.

For example, recoverable errors may not hang your application, but
should be processed by the email notifier. Sending a :once state with
the error description means you can receive an email for each error,
instead of two for entering and exiting the error state.